+++ statistics/handle/autoanalyze/BUILD.bazel
@@ -29,7 +29,7 @@ go_test(
     timeout = "short",
     srcs = ["autoanalyze_test.go"],
     flaky = True,
-    shard_count = 3,
+    shard_count = 6,
     deps = [
         ":autoanalyze",
         "//parser/model",
