+++ statistics/handle/updatetest/BUILD.bazel
@@ -8,7 +8,7 @@ go_test(
         "update_test.go",
     ],
     flaky = True,
-    shard_count = 25,
+    shard_count = 23,
     deps = [
         "//parser/model",
         "//parser/mysql",
