+++ statistics/handle/updatetest/update_test.go
@@ -496,79 +496,6 @@ func TestAutoUpdatePartition(t *testing.T) {
 	})
 }
 
-func TestAutoAnalyzeOnEmptyTable(t *testing.T) {
-	store, dom := testkit.CreateMockStoreAndDomain(t)
-	tk := testkit.NewTestKit(t, store)
-
-	oriStart := tk.MustQuery("select @@tidb_auto_analyze_start_time").Rows()[0][0].(string)
-	oriEnd := tk.MustQuery("select @@tidb_auto_analyze_end_time").Rows()[0][0].(string)
-	defer func() {
-		tk.MustExec(fmt.Sprintf("set global tidb_auto_analyze_start_time='%v'", oriStart))
-		tk.MustExec(fmt.Sprintf("set global tidb_auto_analyze_end_time='%v'", oriEnd))
-	}()
-
-	tt := time.Now().Add(-1 * time.Minute)
-	h, m := tt.Hour(), tt.Minute()
-	start, end := fmt.Sprintf("%02d:%02d +0000", h, m), fmt.Sprintf("%02d:%02d +0000", h, m)
-	tk.MustExec(fmt.Sprintf("set global tidb_auto_analyze_start_time='%v'", start))
-	tk.MustExec(fmt.Sprintf("set global tidb_auto_analyze_end_time='%v'", end))
-	dom.StatsHandle().HandleAutoAnalyze(dom.InfoSchema())
-
-	tk.MustExec("use test")
-	tk.MustExec("create table t (a int, index idx(a))")
-	// to pass the stats.Pseudo check in autoAnalyzeTable
-	tk.MustExec("analyze table t")
-	// to pass the AutoAnalyzeMinCnt check in autoAnalyzeTable
-	tk.MustExec("insert into t values (1)" + strings.Repeat(", (1)", int(autoanalyze.AutoAnalyzeMinCnt)))
-	require.NoError(t, dom.StatsHandle().DumpStatsDeltaToKV(handle.DumpAll))
-	require.NoError(t, dom.StatsHandle().Update(dom.InfoSchema()))
-
-	// test if it will be limited by the time range
-	require.False(t, dom.StatsHandle().HandleAutoAnalyze(dom.InfoSchema()))
-
-	tk.MustExec("set global tidb_auto_analyze_start_time='00:00 +0000'")
-	tk.MustExec("set global tidb_auto_analyze_end_time='23:59 +0000'")
-	require.True(t, dom.StatsHandle().HandleAutoAnalyze(dom.InfoSchema()))
-}
-
-func TestAutoAnalyzeOutOfSpecifiedTime(t *testing.T) {
-	store, dom := testkit.CreateMockStoreAndDomain(t)
-	tk := testkit.NewTestKit(t, store)
-
-	oriStart := tk.MustQuery("select @@tidb_auto_analyze_start_time").Rows()[0][0].(string)
-	oriEnd := tk.MustQuery("select @@tidb_auto_analyze_end_time").Rows()[0][0].(string)
-	defer func() {
-		tk.MustExec(fmt.Sprintf("set global tidb_auto_analyze_start_time='%v'", oriStart))
-		tk.MustExec(fmt.Sprintf("set global tidb_auto_analyze_end_time='%v'", oriEnd))
-	}()
-
-	tt := time.Now().Add(-1 * time.Minute)
-	h, m := tt.Hour(), tt.Minute()
-	start, end := fmt.Sprintf("%02d:%02d +0000", h, m), fmt.Sprintf("%02d:%02d +0000", h, m)
-	tk.MustExec(fmt.Sprintf("set global tidb_auto_analyze_start_time='%v'", start))
-	tk.MustExec(fmt.Sprintf("set global tidb_auto_analyze_end_time='%v'", end))
-	dom.StatsHandle().HandleAutoAnalyze(dom.InfoSchema())
-
-	tk.MustExec("use test")
-	tk.MustExec("create table t (a int)")
-	// to pass the stats.Pseudo check in autoAnalyzeTable
-	tk.MustExec("analyze table t")
-	// to pass the AutoAnalyzeMinCnt check in autoAnalyzeTable
-	tk.MustExec("insert into t values (1)" + strings.Repeat(", (1)", int(autoanalyze.AutoAnalyzeMinCnt)))
-	require.NoError(t, dom.StatsHandle().DumpStatsDeltaToKV(handle.DumpAll))
-	require.NoError(t, dom.StatsHandle().Update(dom.InfoSchema()))
-
-	require.False(t, dom.StatsHandle().HandleAutoAnalyze(dom.InfoSchema()))
-	tk.MustExec("analyze table t")
-
-	tk.MustExec("alter table t add index ia(a)")
-	require.False(t, dom.StatsHandle().HandleAutoAnalyze(dom.InfoSchema()))
-
-	tk.MustExec("set global tidb_auto_analyze_start_time='00:00 +0000'")
-	tk.MustExec("set global tidb_auto_analyze_end_time='23:59 +0000'")
-	require.True(t, dom.StatsHandle().HandleAutoAnalyze(dom.InfoSchema()))
-}
-
 func TestIssue25700(t *testing.T) {
 	store, dom := testkit.CreateMockStoreAndDomain(t)
 	tk := testkit.NewTestKit(t, store)
